Changed distance to float value

+++ LDLSH/src/main/java/SystemLayer/Processes/DistanceMeasurerImpl/DistanceMeasurer.java
@@ -3,5 +3,5 @@ package SystemLayer.Processes.DistanceMeasurerImpl;
 import SystemLayer.Data.DataObjectsImpl.DataObject;
 
 public interface DistanceMeasurer {
-    int getDistance(DataObject object_a, DataObject object_b);
+    float getDistance(DataObject object_a, DataObject object_b);
 }
