+++ LDLSH/src/main/java/SystemLayer/Processes/DistanceMeasurerImpl/JaccardDistance.java
@@ -4,7 +4,7 @@ import SystemLayer.Data.DataObjectsImpl.DataObject;
 
 public class JaccardDistance implements DistanceMeasurer{
     @Override
-    public int getDistance(DataObject object_a, DataObject object_b) {
+    public float getDistance(DataObject object_a, DataObject object_b) {
         return 0;
     }
 }
